feat: integrate ProjectCard component in ProjectsPage with proper type conversion
✨ Features:
- Use reusable ProjectCard component from dashboard in ProjectsPage
- Add type conversion function to map Project to RecentProject format
- Maintain FloatingCircles animations for consistent UX

🔧 Implementation:
- Add convertToRecentProject function to handle type mapping
- Include default values for progress (0%) and dueDate (30 days)
- Calculate team size from memberIds.length
- Map project status to display format (Active/Completed/Inactive)

🎯 Benefits:
- Consistent project card design across dashboard and projects page
- Reusable component reduces code duplication
- Type-safe conversion ensures compatibility
- Enhanced visual consistency with floating animations

📦 Files Updated:
- ProjectsPage.tsx: Added type conversion and ProjectCard usage
- ProjectCard.tsx: Extracted reusable component from dashboard
- RecentProjects.tsx: Updated to maintain compatibility
